@@ -34,7 +34,7 @@ static inline void compilerBarrier() {
 #error "Your compiler is not supported"
 #endif
 
-#if defined(__i386__) || defined(__x86_64__) || defined(__arm__)
+#if defined(__i386__) || defined(__x86_64__) || defined(__arm__) || defined(__aarch64__)
 #  define acquireBarrier() compilerBarrier()
 #  define releaseBarrier() compilerBarrier()
 #else
