@@ -204,7 +204,7 @@ private:
 }
 
 std::shared_ptr<core::posix::SignalTrap> core::posix::trap_signals_for_process(
-        std::initializer_list<core::posix::Signal> blocked_signals)
+	std::initializer_list<core::posix::Signal> blocked_signals)
 {
     return std::make_shared<impl::SignalTrap>(
                 impl::SignalTrap::Scope::process,
