@@ -42,6 +42,19 @@ Platform::Platform(
     : input_manager_(input_manager),
       event_thread_running_(false),
       single_window_(single_window) {
+
+  // Don't block the screensaver from kicking in. It will be blocked
+  // by the desktop shell already and we don't have to do this again.
+  // If we would leave this enabled it will prevent systems from
+  // suspending correctly.
+  SDL_SetHint(SDL_HINT_VIDEO_ALLOW_SCREENSAVER, "1");
+
+#ifdef SDL_HINT_VIDEO_X11_NET_WM_BYPASS_COMPOSITOR
+  // Don't disable compositing
+  // Available since SDL 2.0.8
+  SDL_SetHint(SDL_HINT_VIDEO_X11_NET_WM_BYPASS_COMPOSITOR, "0");
+#endif
+
   if (SDL_Init(SDL_INIT_VIDEO | SDL_INIT_AUDIO | SDL_INIT_EVENTS) < 0) {
     const auto message = utils::string_format("Failed to initialize SDL: %s", SDL_GetError());
     BOOST_THROW_EXCEPTION(std::runtime_error(message));
@@ -70,6 +83,7 @@ Platform::Platform(
   }
 
   graphics::emugl::DisplayInfo::get()->set_resolution(display_frame.width(), display_frame.height());
+  display_frame_ = display_frame;
 
   pointer_ = input_manager->create_device();
   pointer_->set_name("anbox-pointer");
@@ -92,6 +106,30 @@ Platform::Platform(
   keyboard_->set_physical_location("none");
   keyboard_->set_key_bit(BTN_MISC);
   keyboard_->set_key_bit(KEY_OK);
+
+#ifdef ENABLE_TOUCH_INPUT
+  touch_ = input_manager->create_device();
+  touch_->set_name("anbox-touch");
+  touch_->set_driver_version(1);
+  touch_->set_input_id({BUS_VIRTUAL, 4, 4, 4});
+  touch_->set_physical_location("none");
+  touch_->set_key_bit(BTN_TOUCH);
+  touch_->set_key_bit(BTN_TOOL_FINGER);
+
+  touch_->set_abs_bit(ABS_MT_SLOT);
+  touch_->set_abs_max(ABS_MT_SLOT, 10);
+  touch_->set_abs_bit(ABS_MT_TOUCH_MAJOR);
+  touch_->set_abs_max(ABS_MT_TOUCH_MAJOR, 127);
+  touch_->set_abs_bit(ABS_MT_TOUCH_MINOR);
+  touch_->set_abs_max(ABS_MT_TOUCH_MINOR, 127);
+  touch_->set_abs_bit(ABS_MT_POSITION_X);
+  touch_->set_abs_max(ABS_MT_POSITION_X, display_frame.width());
+  touch_->set_abs_bit(ABS_MT_POSITION_Y);
+  touch_->set_abs_max(ABS_MT_POSITION_Y, display_frame.height());
+  touch_->set_abs_bit(ABS_MT_TRACKING_ID);
+  touch_->set_abs_max(ABS_MT_TRACKING_ID, 10);
+  touch_->set_prop_bit(INPUT_PROP_DIRECT);
+#endif
 
   event_thread_ = std::thread(&Platform::process_events, this);
 }
@@ -130,12 +168,18 @@ void Platform::process_events() {
             }
           }
           break;
+        case SDL_KEYDOWN:
+        case SDL_KEYUP:
+          if (keyboard_)
+            process_input_event(event);
+          break;
         case SDL_MOUSEMOTION:
         case SDL_MOUSEBUTTONDOWN:
         case SDL_MOUSEBUTTONUP:
         case SDL_MOUSEWHEEL:
-        case SDL_KEYDOWN:
-        case SDL_KEYUP:
+        case SDL_FINGERDOWN:
+        case SDL_FINGERUP:
+        case SDL_FINGERMOTION:
           process_input_event(event);
           break;
         default:
@@ -148,25 +192,26 @@ void Platform::process_events() {
 void Platform::process_input_event(const SDL_Event &event) {
   std::vector<input::Event> mouse_events;
   std::vector<input::Event> keyboard_events;
+  std::vector<input::Event> touch_events;
 
   std::int32_t x = 0;
   std::int32_t y = 0;
+
   SDL_Window *window = nullptr;
 
   switch (event.type) {
+    // Mouse
     case SDL_MOUSEBUTTONDOWN:
       mouse_events.push_back({EV_KEY, BTN_LEFT, 1});
-      mouse_events.push_back({EV_SYN, SYN_REPORT, 0});
       break;
     case SDL_MOUSEBUTTONUP:
       mouse_events.push_back({EV_KEY, BTN_LEFT, 0});
-      mouse_events.push_back({EV_SYN, SYN_REPORT, 0});
       break;
     case SDL_MOUSEMOTION:
       if (!single_window_) {
         // As we get only absolute coordindates relative to our window we have to
         // calculate the correct position based on the current focused window
-        window = SDL_GetWindowFromID(event.window.windowID);
+        window = SDL_GetWindowFromID(focused_sdl_window_id_);
         if (!window) break;
 
         SDL_GetWindowPosition(window, &x, &y);
@@ -190,12 +235,12 @@ void Platform::process_input_event(const SDL_Event &event) {
       // was moved. They are not used to find out the exact position.
       mouse_events.push_back({EV_REL, REL_X, event.motion.xrel});
       mouse_events.push_back({EV_REL, REL_Y, event.motion.yrel});
-      mouse_events.push_back({EV_SYN, SYN_REPORT, 0});
       break;
     case SDL_MOUSEWHEEL:
       mouse_events.push_back(
           {EV_REL, REL_WHEEL, static_cast<std::int32_t>(event.wheel.y)});
       break;
+    // Keyboard
     case SDL_KEYDOWN: {
       const auto code = KeycodeConverter::convert(event.key.keysym.scancode);
       if (code == KEY_RESERVED) break;
@@ -208,13 +253,109 @@ void Platform::process_input_event(const SDL_Event &event) {
       keyboard_events.push_back({EV_KEY, code, 0});
       break;
     }
+#ifdef ENABLE_TOUCH_INPUT
+    // Touch screen
+    case SDL_FINGERDOWN: {
+      touch_events.push_back({EV_ABS, ABS_MT_TRACKING_ID, static_cast<std::int32_t>(event.tfinger.fingerId)});
+      touch_events.push_back({EV_ABS, ABS_MT_SLOT, 0});
+      touch_events.push_back({EV_KEY, BTN_TOUCH, 1});
+      touch_events.push_back({EV_KEY, BTN_TOOL_FINGER, 1});
+
+      if (!calculate_touch_coordinates(event, x, y))
+        break;
+
+      touch_events.push_back({EV_ABS, ABS_MT_POSITION_X, x});
+      touch_events.push_back({EV_ABS, ABS_MT_POSITION_Y, y});
+
+      touch_events.push_back({EV_ABS, ABS_MT_TOUCH_MAJOR, 24});
+      touch_events.push_back({EV_ABS, ABS_MT_TOUCH_MINOR, 24});
+
+      touch_events.push_back({EV_SYN, SYN_REPORT, 0});
+      break;
+    }
+	  case SDL_FINGERUP: {
+      touch_events.push_back({EV_ABS, ABS_MT_TRACKING_ID, -1});
+      touch_events.push_back({EV_ABS, ABS_MT_SLOT, 0});
+      touch_events.push_back({EV_KEY, BTN_TOUCH, 0});
+      touch_events.push_back({EV_KEY, BTN_TOOL_FINGER, 0});
+      touch_events.push_back({EV_SYN, SYN_REPORT, 0});
+      break;
+    }
+	  case SDL_FINGERMOTION: {
+      touch_events.push_back({EV_ABS, ABS_MT_SLOT, 0});
+
+      if (!calculate_touch_coordinates(event, x, y))
+        break;
+
+      touch_events.push_back({EV_ABS, ABS_MT_POSITION_X, x});
+      touch_events.push_back({EV_ABS, ABS_MT_POSITION_Y, y});
+      DEBUG("ABS_MT_POSITION: x: %i y: %i", x,y);
+
+      touch_events.push_back({EV_ABS, ABS_MT_TOUCH_MAJOR, 24});
+      touch_events.push_back({EV_ABS, ABS_MT_TOUCH_MINOR, 24});
+      touch_events.push_back({EV_SYN, SYN_REPORT, 0});
+      break;
+    }
+#endif
     default:
       break;
   }
 
-  if (mouse_events.size() > 0) pointer_->send_events(mouse_events);
+  if (mouse_events.size() > 0) {
+    mouse_events.push_back({EV_SYN, SYN_REPORT, 0});      
+    pointer_->send_events(mouse_events);
+  }
 
-  if (keyboard_events.size() > 0) keyboard_->send_events(keyboard_events);
+  if (keyboard_events.size() > 0)
+    keyboard_->send_events(keyboard_events);
+
+#ifdef ENABLE_TOUCH_INPUT
+  if (touch_events.size() > 0)
+    touch_->send_events(touch_events);
+#endif
+}
+
+bool Platform::calculate_touch_coordinates(const SDL_Event &event,
+                                           std::int32_t &x,
+                                           std::int32_t &y) {
+  std::int32_t rel_x = 0;
+  std::int32_t rel_y = 0;
+
+  SDL_Window *window = nullptr;
+
+  window = SDL_GetWindowFromID(focused_sdl_window_id_);
+  // before SDL 2.0.7 on X11 tfinger coordinates are not normalized
+  if (!SDL_VERSION_ATLEAST(2,0,7) && (event.tfinger.x > 1 || event.tfinger.y > 1)) {
+    rel_x = event.tfinger.x;
+    rel_y = event.tfinger.y;
+  } else {
+    if (window) {
+      SDL_GetWindowSize(window, &rel_x, &rel_y);
+      rel_x *= event.tfinger.x;
+      rel_y *= event.tfinger.y;
+    } else {
+      rel_x = display_frame_.width() * event.tfinger.x;
+      rel_y = display_frame_.height() * event.tfinger.y;
+    }
+  }
+
+  if (!single_window_) {
+    if (!window) {
+      return false;
+    }
+    // As we get only absolute coordindates relative to our window we have to
+    // calculate the correct position based on the current focused window
+    SDL_GetWindowPosition(window, &x, &y);
+    x += rel_x;
+    y += rel_y;
+  } else {
+    // When running the whole Android system in a single window we don't
+    // need to reacalculate and the pointer position as they are already
+    // relative to our window.
+    x = rel_x;
+    y = rel_y;
+  }
+  return true;
 }
 
 Window::Id Platform::next_window_id() {
@@ -231,6 +372,7 @@ std::shared_ptr<wm::Window> Platform::create_window(
 
   auto id = next_window_id();
   auto w = std::make_shared<Window>(renderer_, id, task, shared_from_this(), frame, title, !window_size_immutable_);
+  focused_sdl_window_id_ = w->window_id();
   windows_.insert({id, w});
   return w;
 }
@@ -250,8 +392,10 @@ void Platform::window_wants_focus(const Window::Id &id) {
   auto w = windows_.find(id);
   if (w == windows_.end()) return;
 
-  if (auto window = w->second.lock())
+  if (auto window = w->second.lock()) {
+    focused_sdl_window_id_ = window->window_id();
     window_manager_->set_focused_task(window->task());
+  }
 }
 
 void Platform::window_moved(const Window::Id &id, const std::int32_t &x,
