@@ -35,10 +35,8 @@ void BootPropertiesMessageProcessor::handle_command(
 void BootPropertiesMessageProcessor::list_properties() {
   std::vector<std::string> properties = {
       // TODO(morphis): Using HDPI here for now but should be adjusted to the
-      // device
-      // we're running on.
-      utils::string_format("ro.sf.lcd_density=%d",
-                           static_cast<int>(graphics::DensityType::medium)),
+      // device we're running on
+      utils::string_format("ro.sf.lcd_density=%d", static_cast<int>(graphics::current_density())),
   };
 
   for (const auto &prop : properties) {
