@@ -48,10 +48,9 @@ std::string read_file_if_exists_or_throw(const std::string &file_path) {
 
   std::ifstream file;
   file.open(file_path, std::ifstream::in);
-  std::string content;
-  file >> content;
-  file.close();
-  return content;
+  std::stringstream buffer;
+  buffer << file.rdbuf();
+  return buffer.str();
 }
 
 bool write_to_file(const std::string &file_path, const std::string &content) {
